Split routes into separate files
Split the file 'routes.py' into separate files in the 'routes' folder
because the file has grown too large. Replace references to old routes
with references to new routes. Rearrange 'app.py' so that it can load
new routes of type 'Blueprint'. As part of this arrangement, move 'db'
to its own module.